% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/completeMatrixFunction.R
\name{completeMatrix}
\alias{completeMatrix}
\title{Impute missing values in a matrix using an expectation maximization approach based on lasso regression models.}
\usage{
completeMatrix(allIc50, nPerms = 50)
}
\arguments{
\item{allIc50}{A matrix of drug sensitivity data with missing values,}

\item{nPerms}{The number of iterations that the EM-algorithm should run.}
}
\description{
Impute missing values in a matrix using an expectation maximization approach based on lasso regression models.
}

