% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gldsCorrectedAssoc.R
\name{gldsCorrectedAssoc}
\alias{gldsCorrectedAssoc}
\title{Calculate assocations GLDS corrected assocations between drug sensitivity data and some predictive marker (e.g. somatic mutations)}
\usage{
gldsCorrectedAssoc(drugMat, drugRelationshipList, markerMat,
  numCorDrugsExclude = 100, minMuts = 5, additionalCovariateMatrix = NULL)
}
\arguments{
\item{drugMat}{Matrix of drug sensitivity data, with drug names as column names and samples (e.g. cell lines) as row names.}

\item{drugRelationshipList}{A list, which for each drug contains a vector of 1's and 0's, indicating whether the drugs are related. I.e. if both drugs were an inhibitor of ERBB2, that position would contain a 1. This order of the drugs should be the same as the order of the drug sensitivty matrix, i.e. drugMat.}

\item{markerMat}{A matrix containing the data for which you are looking for an association with drug sensitivity, i.e. a matrix of somatic mutation status. This matrix has samples names (i.e. cell lines) as columns and marker names (e.g. gene names) as row names.}

\item{numCorDrugsExclude}{When calculating GLDS, one of the steps involves searching for drugs with a high correlation with the drug on interest and excluding them from the calculation of GLDS (so as to not remove drug specific signal). As a rule of thumb, about 25% of the size of the dataset may be suitable.}

\item{minMuts}{For each marker, i.e. row of markerMat, what is the minimum number of non-zero enteries required so that we calculate a P-value. E.g. how many somatic mutations need to be present.}

\item{additionalCovariateMatrix}{A matrix containing additional covariates to be fit in the drug biomarker association models. This could be, for example, tissue of origin or cancer type. Columns should be sample names (i.e. cell lines) and rows should be the row names.}
}
\description{
Calculate assocations GLDS corrected assocations between drug sensitivity data and some predictive marker (e.g. somatic mutations)
}

